Use cluster custom header for DisableSelfContainedVHD
